Refactor configurations of ResourceSystem
* Unify column names, table names, index names and type declarations
  in configurations of entities in ResourceSystem
* Add default value for audit fields CreatedAt and UpdatedAt

+++ src/Domain/Entities/ResourceSystem/AmusementRide.cs
@@ -41,12 +41,12 @@ public class AmusementRide
     /// <summary>
     /// 容量
     /// </summary>
-    public short Capacity { get; set; }
+    public int Capacity { get; set; }
 
     /// <summary>
     /// 单趟时间(秒)
     /// </summary>
-    public short Duration { get; set; }
+    public int Duration { get; set; }
 
     /// <summary>
     /// 最低身高限制(cm)
